Rework node drivers to expose predictable node names
It was deciced that a 4DO-3DI-3AI with serial number
26 should be controlled using a REST-URLs like this:

 http://iobus-server/nodes/4DO-3DI-3AI-00005.00026/pins/

This URL should also be alias-free, meaning the node may
not be reached using other URLs.

This commit implements this by changing the way node
names are generated.
It also re-works the individual node drivers so that
they fit on a screen page and renames them to reflect
the official (not engineerin) product names.