Use the locator sdo as ping target
There has to be a period check if devices on the bus are still
responding. This check was performed using a dummy sdo read request
and its actual response was ignored.

Replace this dummy read by an actual read of the locator status
so that devices with an active locator can be highlighted in the
node list.