Disable flashing of custom firmware by default
The feature to flash own firmware by default is intended for a
development use-case where the operator is close to the device
or is aware of the risk to brick a device on a remote location.

This feature should only be enabled if really needed.
This change adds a new command line switch --allow-custom-firmware
that enables the upload and flash feature for arbitrary firmware.
The switch defaults to False thus disabling the feature by default.

The new switch is then plumbed through the server into the frontend.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>