logging/aiohttp: Only enable access when DEBUG-log is enabled
Per default access-logging is done with log-level INFO.
Tools like labgrid tend to poll the REST-API of this server. This leads
to a lot of noise in the log if just INFO-Log is enabled.

This change disables the access-log in all log-levels but DEBUG.

This adds to the objective to make the INFO-log more informative for the
user: INFO-log should contain messages that carry information that is
relevant for the user - and not for the developer.
(A developer can still use the DEBUG-log level to gain full insight into
whats going on.)