logging/server: log actions where user interats with a node
With this change every access (via REST or Web-Interface) that changes
or accesses the state of a connected node is logged into the INFO log.

This adds to the objective to make the INFO-log more informative for the
user: INFO-log should contain messages that carry information that is
relevant for the user - and not for the developer.
(A developer can still use the DEBUG-log level to gain full insight into
whats going on.)