logging/server: Log wrong user input as warn not error
Until now wrong user input was logged using 'logger.exception()' what
lead to an ERROR and a python stacktrace in the log.
This behavior is problematic in two ways

a) Wrong user input (e.g. request pins from an unknown node) is
   recoverable and thus not an ERROR.
   Since this information is relevant for the user I would still
   log such event as warning.
b) Logging stack traces for an expected exception due to user input:
   We have to anticipate that a user will provide us with wrong input.
   In such case a stack trace does not provide any additional
   information to the user.
   Stack traces are a tool for the developer and thus only unexpected
   exceptions should log a stack trace (that the user can use to either
   debug the cause or at least report the issue).

This commit adds better exception handling and logging to the methods of
the REST-API.
Also: With this change the REST-API now yields a useful error message if
a pin is unknown. Until now only a KeyError was reported.