Show LSS state and CAN TX error on status page
This change propagates the network state into the ``/server-info/`` API
and displays this information on the status page.

Without this change the user did not get any feedback that a scan is in
progress. Especially with a new node (where the LSS address is not yet
cached) this takes a few seconds.
With this change a user gets feedback once a new node is discovered on
the bus and LSS tries to determine it's LSS address.

While implementing this I also plumbed the CAN TX error from the network
into the fronted.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>
Co-authored-by: Jan Lübbe <jlu@pengutronix.de>