Remove use of deprecated (in version 4.0) python-can API
Some time between python-can version 2.0 and 3.0 the extended_id parameter of
the Message class was renamed to is_extended_id but the old name was kept for
backward compatibility as a deprecated parameter.

With the release of python-can 4.0 this parameter was removed[1].

Change the parameter name to make the server python-can 4.0 compatible.

[1]: https://github.com/hardbyte/python-can/commit/6a38debd041e4448017e451517b15daaa0933dd5

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>