lxa_iobus: move the project from setup.py to pyproject.toml
The pyproject.toml format is the new and modern way to package python
projects. We still use setuptools under the hood to do the packaging,
but configure it in the pyproject.toml.

This change also moves the executable scripts from the bin folder,
where they were e.g. not caught by ruff by default,
into the python project folder itself.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>