lxa_iobus: lpc11xxcanisp: make methods in CanIsp async
The CanIsp class has gone through a few iterations as a commandline
tool using the external canopen library before it was included in the
server and used the lxa_iobus intenal async canopen implementation.
This is why it stayed non-async until now, making integration with the
rest of the IOBus server cumbersome.

Make CanIsp async-native.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>