lxa_iobus: server: add API endpoints to get/set raw SDOs on the node
Add support for talking to IOBus nodes connected to an IOBus server
directly via SDO upload/download via HTTP.

This is an obvious abstraction layer violation and not intended for
external consumers like the web interface or labgrid.
External consumers should be provided with useful abstractions for
node features, instead of having to assemble and parse raw SDO messages.
These endpoints are intended as a development utility that can come
in handy when e.g. developing a new kind of device where it is not yet
clear what a proper API should look like.
Having these raw interfaces allows writing a commandline util that
talks directly to the node first and only later specifying an API,
which is arguably better than coming up with an ad-hoc API that
one regrets later.

Speaking of APIs that one regrets later, these new endpoints are
placed behind a new `/api/v2/` URL prefix.
This should make it easier than now to iteratively improve the
API using versions and backwards compatibility shims in the server.

Right now these endpoints are the only ones behind the `/api/v2`
prefix, but we can add others later, that fix errors we made in the
(implicit) v1 of the API, like:

  - Not using HTTP status codes (other than 200)
  - Exporting numbers as pre-formatted strings
  - Not using correct content-types for requests and responses.
    (E.g. not using text/event-stream as request content-type
    for endpoints where the client wants to be notified about
    changes and instead using a full-blown RPC library).

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>