lxa_iobus: node: replace ad-hoc struct parsing with central object_directory
Previously unpacking and packing structs to send to the node was always
done ad-hoc, where it was needed.
This required a deep knowledge of the different IOBus protocols whenever
any of the node handling code should be changed and made it difficult
to use lxa_iobus as a library outside of the server, because there was
no complete documentation of IOBus CANopen endpoints.

Centralize the IOBus protocol handling in a single object_directory.py
file.

This also changes the way IOBus protocols on nodes are enumerated.
Now any node that advertises support for a specific protocol will appear
in the web interface with a list of inputs/outputs/adcs it has and
not just as an empty dummy node.

This change also removes the notion of a "Node Driver".
A node is just a collection of the IOBus protocols / CANopen endpoints
it supports. The drivers only served to give the inputs and outputs
and the node itself human readable and descriptive names.

This is now done in a list of products in products.py.
A node can however be perfectly usable without appearing in the products
list, and will just use default names for all inputs and outputs.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>