lxa_iobus: network: do not add new nodes before they are set up
Previously a node could be accessed via e.g. the API before it was
fully set up (e.g. before the supported protocols were enumerated).
This looked a bit weird in the web interface but did not do serious
harm because enumartion usually happened quite fast.

Completely prevent the half-initialized nodes from appearing in the web
interface so that the presented state is always consistent,
by only adding the node to the list once it is fully set up.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>