lxa_iobus: network: initialize attributes in __init__
This improves the usecase where the Network class should be used outside
of the server (e.g. in the flashing commandline utils or when using
lxa_iobus as a library), because it means that e.g. await_interface_is_up()
can be used before run() got around to setting the _interface_state.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>