lxa_iobus: network: make {load,write}_lss_address_cache sync
Filesystem access is generally a blocking operation and some low latency
or high performance software needs to take that into account.
We are neither. The single connect client can likely accept the
millisecond latency spike when the lss cache file is written to
disk whenever a not previously seen node is connected for the first time.

The `if True:` is only there to separate the formatting change from
the code change and will be removed in a followup commit.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>