lxa_iobus: use explicit little ending struct packing/unpacking everywhere
When specifying a struct format without a "@,=,<,>!" prefix python will
default to native struct encoding with native byte order, alignment and
sizes. Which is what we want if the native byte order is little endian,
the alignment is "close enough" and the sizes are not crazy.

This did not seam to affect anyone yet, because we do not rely on alignment
to much and most systems are little endian and not crazy.

But we should not rely on that and explicity state the format we want
so the one person with a big endian MIPS box can use the IOBus server.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>