lxa_iobus: server: fix string/integer type confusion in set_pin()
The form encoded parameter `value` sent via POST request is parsed
by aiohttp so it can be used in set_pin().
It is however passed in as string ("0", "1" or "toggle") so it will
always be considered true-ish, resulting in the current implementation
always turning the output on, even if "0" was requested.

Treat the value as integer instead (as long as it is not "toggle")
and check that for true-ish-ness.

Fixes: 9236c1d ("lxa_iobus: node: replace ad-hoc struct parsing ...
Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>