pyproject.toml: use setuptools_scm to determine version number
This means we no longer have to update the version number manually
when creating a new release because just setting a tag suffices.

It does however also add some complexity and magic to the project,
like incorrect version numbers when the git history is not complete
(hence why we now clone the whole history in the build job) and
all files included in the git repository now being shipped in the
source distribution (while previously only files mentioned in the
MANIFEST.in were included).

The beneftis when it comes to automating the release process do
however outweight the drawbacks of added compexity.

We use `local_scheme = "no-local-version"` to prevent random files
in the project directory from causing "+dirty" version numbers.

Thus use setuptools_scm to determine the version number from
git tags.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>