node: optick: fix capture output with multiple outputs
The raw events are lists with timestamps and state per input.
The output contains the state of all the outputs for each timstamp.

The lists must be merged and sorted before adding the state of the other
inputs for each timestamp. Otherwise an old state value is used and the
output is incorrect.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>